Added file. See comments at in source code
This is motivated by the fact that `x =  MATLAnswer.fetch();` gives an error, apparently because there are too many answers and some download limitation kicks in. Not sure if that can be fixed, maybe using some pause to comply with the limitation. So I've written this file, which downloads answers after a specified date and merges them with answers contained in a specified file (which was generated in the specified date), removing duplicated that may exist in the specified date.